+++ src/commands.ts
@@ -27,6 +27,7 @@ export * from './commands/diffWithWorking';
 export * from './commands/externalDiff';
 export * from './commands/gitCommands';
 export * from './commands/inviteToLiveShare';
+export * from './commands/logging';
 export * from './commands/openAssociatedPullRequestOnRemote';
 export * from './commands/openBranchesOnRemote';
 export * from './commands/openBranchOnRemote';