+++ src/commands/common.ts
@@ -70,6 +70,8 @@ export enum Commands {
 	DiffWithWorkingInDiffRight = 'gitlens.diffWithWorkingInDiffRight',
 	DiffLineWithWorking = 'gitlens.diffLineWithWorking',
 	DisconnectRemoteProvider = 'gitlens.disconnectRemoteProvider',
+	DisableDebugLogging = 'gitlens.disableDebugLogging',
+	EnableDebugLogging = 'gitlens.enableDebugLogging',
 	DisableRebaseEditor = 'gitlens.disableRebaseEditor',
 	EnableRebaseEditor = 'gitlens.enableRebaseEditor',
 	ExternalDiff = 'gitlens.externalDiff',